+++ graph_harness/bootstrap/__init__.py
@@ -1,0 +1,12 @@
+from .preflight import PreflightFinding, PreflightResult, run_preflight
+from .scaffold import BootstrapAction, BootstrapPlan, apply_bootstrap, plan_bootstrap
+
+__all__ = [
+    "BootstrapAction",
+    "BootstrapPlan",
+    "PreflightFinding",
+    "PreflightResult",
+    "apply_bootstrap",
+    "plan_bootstrap",
+    "run_preflight",
+]